# openapi/openapi.yaml
---
openapi: 3.1.0
components:
  schemas:
    CreateSphereRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          pattern: \S
    CreateSphereResponse:
      type: object
      properties:
        sphereId:
          type: integer
          format: int64
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          $ref: "#/components/schemas/Instant"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
    ListSphereResponse:
      type: object
      properties:
        spheres:
          type: array
          items:
            $ref: "#/components/schemas/SphereListInfoDTO"
    NodeDownloadUrlResponse:
      type: object
      properties:
        downloadUrl:
          type: string
    NodeTreeDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        parentId:
          type: integer
          format: int64
        name:
          type: string
        type:
          $ref: "#/components/schemas/NodeType"
    NodeTreeResponse:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/NodeTreeDTO"
    NodeType:
      type: string
      enum:
      - FILE
      - FOLDER
    SphereListInfoDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        ownerId:
          type: integer
          format: int64
        name:
          type: string
  securitySchemes:
    SecurityScheme:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
      summary: Login User
      tags:
      - Authenthication Resource
  /auth/logout:
    post:
      parameters:
      - name: refreshToken
        in: cookie
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Logout User
      tags:
      - Authenthication Resource
  /auth/refresh:
    post:
      parameters:
      - name: refreshToken
        in: cookie
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
      summary: Refresh
      tags:
      - Authenthication Resource
  /auth/signup:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        "200":
          description: Signup successful
          content:
            application/json:
              schema:
                type: object
        "409":
          description: Email or username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
      summary: Signup User
      tags:
      - Authenthication Resource
  /hello:
    get:
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
      summary: Hello
      tags:
      - Greeting Resource
  /node:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        "201":
          description: Nodes created successfully
          content:
            application/json:
              schema:
                type: object
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create Node
      tags:
      - Node Resource
      security:
      - SecurityScheme: []
  /node/{id}:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Node Data
      tags:
      - Node Resource
      security:
      - SecurityScheme: []
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Delete Node
      tags:
      - Node Resource
      security:
      - SecurityScheme: []
  /node/{id}/downloadUrl:
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Download url got succesfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDownloadUrlResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Node Download Url
      tags:
      - Node Resource
      security:
      - SecurityScheme: []
  /sphere:
    get:
      parameters:
      - name: ownerId
        in: query
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Spheres retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListSphereResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Spheres
      tags:
      - Sphere Resource
      security:
      - SecurityScheme: []
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSphereRequest"
        required: true
      responses:
        "201":
          description: Sphere created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateSphereResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
        "400":
          description: Bad Request
      summary: Create Sphere
      tags:
      - Sphere Resource
      security:
      - SecurityScheme: []
  /sphere/{sphereId}/node-tree:
    get:
      parameters:
      - name: sphereId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Node tree retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeTreeResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Get Node Tree By Sphere Id
      tags:
      - Sphere Resource
      security:
      - SecurityScheme: []
info:
  title: allround-backend API
  version: 1.0.0-SNAPSHOT
servers:
- url: http://localhost:8080
  description: Auto generated value
- url: http://0.0.0.0:8080
  description: Auto generated value